Fix some warnings in GUI part

+++ GUI/Dialogs4/CalcCheckDialog.cpp
@@ -35,7 +35,7 @@ CalcCheckDialog::CalcCheckDialog(QWidget* parent,
    setWindowTitle( "Check boxes Calculator" );
 
    if( nO >= 0 )
-        rObj = &aObj[nO];
+        rObj = &aObj[static_cast<size_t>(nO)];
    else
         rObj = new TObject( "tempor", ftCheckBox, 1, 1, false, 'N', "temporary object");
