+++ GUI/Dialogs4/SelectDialog.cpp
@@ -45,7 +45,7 @@ SelectDialog::SelectDialog( QWidget* parent, const char* title,
     }
 
     pList->setSelectionMode(QAbstractItemView::SingleSelection); //pList->setMultiSelection(false);
-    if( sel < 0 || (uint)sel > list.GetCount() )
+    if( sel < 0 || sel > list.GetCount() )
 	    sel = 0;
     pList->setCurrentRow(sel); // setSelected(sel, true);
 
@@ -77,7 +77,7 @@ SelectDialog::SelectDialog( QWidget* parent, const char* title,
     }
 
     pList->setSelectionMode(QAbstractItemView::SingleSelection); //pList->setMultiSelection(false);
-    if( sel < 0 || (uint)sel > list.GetCount() )
+    if( sel < 0 || sel > list.GetCount() )
  	    sel = 0;
      pList->setCurrentRow(sel); // setSelected(sel, true);
