+++ GUI/Services4/service.h
@@ -158,7 +158,7 @@ vfReacDCSet(QWidget* par, const char * p_key,
 
 bool
 vfRTparmSet(QWidget* par, const char * p_key,
-            char flgs[10], int sizes[7], float val[6],
+            char flgs[10], int sizes[7], double val[6],
             gstring& script, gstring& xName, gstring& yName, TCStringArray& names  );
 
 bool
