+++ Modules/filters_data.h
@@ -111,7 +111,7 @@ struct elmWindowData
          pos1 += 13;
          size_t pos2 = strBuf.find( ";", pos1 );
          if( pos2-pos1 >= 14 )
-           for(int ii=0; ii<14; ii++ )
+           for(size_t ii=0; ii<14; ii++ )
              flags[ii] = (strBuf[ii+pos1] =='+');
       }
   }
@@ -394,7 +394,7 @@ struct setFiltersData
 
 struct tableSetupData
 {
-   int nWin;     // page number
+   size_t nWin;     // page number
    int nObj;     // number of object
    int nIdx;     // line into object
    double val;
